requirements.txt
aioredis==1.2.0
amqp==2.4.2
asgiref==2.3.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
billiard==3.5.0.5
celery==4.2.1
channels==2.1.7
channels-redis==2.3.2
constantly==15.1.0
daphne==2.2.5
Django==2.1.7
django-debug-toolbar==1.11
django-environ==0.4.5
django-extensions==2.1.4
django-model-utils==3.1.2
hiredis==1.0.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
kombu==4.4.0
msgpack==0.5.6
pkg-resources==0.0.0
PyHamcrest==1.9.0
pytz==2018.9
redis==3.2.1
six==1.12.0
sqlparse==0.3.0
Twisted==18.9.0
txaio==18.8.1
vine==1.3.0
zope.interface==4.6.0
